```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/uniseg
                          │  .old.txt   │              .new.txt               │
                          │   sec/op    │   sec/op     vs base                │
SentenceFunctionBytes-10    3.679µ ± 0%   2.209µ ± 0%  -39.93% (p=0.000 n=10)
SentenceFunctionString-10   3.696µ ± 0%   2.165µ ± 0%  -41.43% (p=0.000 n=10)
geomean                     3.687µ        2.187µ       -40.69%

                          │   .old.txt   │              .new.txt               │
                          │     B/op     │    B/op     vs base                 │
SentenceFunctionBytes-10    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
SentenceFunctionString-10   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                          │   .old.txt   │              .new.txt               │
                          │  allocs/op   │ allocs/op   vs base                 │
SentenceFunctionBytes-10    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
SentenceFunctionString-10   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```