@shogo82148 2023-07-24
Before:

![profile001](https://github.com/shogo82148/uniseg/assets/1157344/e2dc65a1-d761-4b45-9719-1d04cad73cea)

After:

![profile002](https://github.com/shogo82148/uniseg/assets/1157344/6c04a762-8e33-42e8-8a06-10aa52df94a2)